--- .vscode/settings.json
@@ -1,9 +1,0 @@
-{
-  "eslint.validate": ["javascript", "html"],
-  "eslint.alwaysShowStatus": true,
-  "editor.tabSize": 2,
-  "git.ignoreLimitWarning": true,
-  "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": true
-  },
-}
