+++ team_first_project01/config/corsConfig.json
