+++ team_first_project01/dao/cartDao.js
@@ -1,11 +1,11 @@
 const { Op } = require('sequelize');
-const { ProductCategory } = require('../models/index');
+const { Cart, Product } = require('../models/index');
 
 const dao = {
   // 등록
   insert(params) {
     return new Promise((resolve, reject) => {
-      ProductCategory.create(params).then((inserted) => {
+      Cart.create(params).then((inserted) => {
         resolve(inserted);
       }).catch((err) => {
         reject(err);
@@ -16,19 +16,26 @@ const dao = {
   selectList(params) {
     // where 검색 조건
     const setQuery = {};
-    if (params.categoryCode) {
+    if (params.cartNumber) {
       setQuery.where = {
         ...setQuery.where,
-        categoryCode: { [Op.like]: `%${params.categoryCode}%` }, // like검색
+        cartNumber: { [Op.like]: `%${params.cartNumber}%` }, // like검색
       };
     }
 
     // order by 정렬 조건
-    setQuery.order = [['categoryCode', 'DESC']];
+    setQuery.order = [['id', 'DESC']];
 
     return new Promise((resolve, reject) => {
-      ProductCategory.findAndCountAll({
+      Cart.findAndCountAll({
         ...setQuery,
+        attributes: { exclude: ['password'] }, // password 필드 제외
+        include: [
+          {
+            model: Product,
+            as: 'Product',
+          },
+        ],
       }).then((selectedList) => {
         resolve(selectedList);
       }).catch((err) => {
@@ -39,8 +46,8 @@ const dao = {
   // 상세정보 조회
   selectInfo(params) {
     return new Promise((resolve, reject) => {
-      ProductCategory.findByPk(
-        params.categoryCode,
+      Cart.findByPk(
+        params.id,
       ).then((selectedInfo) => {
         resolve(selectedInfo);
       }).catch((err) => {
@@ -51,10 +58,10 @@ const dao = {
   // 수정
   update(params) {
     return new Promise((resolve, reject) => {
-      ProductCategory.update(
+      Cart.update(
         params,
         {
-          where: { categoryCode: params.categoryCode },
+          where: { id: params.id },
         },
       ).then(([updated]) => {
         resolve({ updatedCount: updated });
@@ -66,8 +73,8 @@ const dao = {
   // 삭제
   delete(params) {
     return new Promise((resolve, reject) => {
-      ProductCategory.destroy({
-        where: { categoryCode: params.categoryCode },
+      Cart.destroy({
+        where: { id: params.id },
       }).then((deleted) => {
         resolve({ deletedCount: deleted });
       }).catch((err) => {
