+++ team_first_project01/dao/productDao.js
@@ -1,0 +1,87 @@
+const { Op } = require('sequelize');
+const { ProductCategory, Product } = require('../models/index');
+
+const dao = {
+  // 등록
+  insert(params) {
+    return new Promise((resolve, reject) => {
+      Product.create(params).then((inserted) => {
+        resolve(inserted);
+      }).catch((err) => {
+        reject(err);
+      });
+    });
+  },
+  // 리스트 조회
+  selectList(params) {
+    // where 검색 조건
+    const setQuery = {};
+    if (params.productName) {
+      setQuery.where = {
+        ...setQuery.where,
+        productName: { [Op.like]: `%${params.productName}%` }, // like검색
+      };
+    }
+
+    // order by 정렬 조건
+    setQuery.order = [['id', 'DESC']];
+
+    return new Promise((resolve, reject) => {
+      Product.findAndCountAll({
+        ...setQuery,
+        attributes: { exclude: ['password'] }, // password 필드 제외
+        include: [
+          {
+            model: ProductCategory,
+            as: 'ProductCategory',
+          },
+        ],
+      }).then((selectedList) => {
+        resolve(selectedList);
+      }).catch((err) => {
+        reject(err);
+      });
+    });
+  },
+  // 상세정보 조회
+  selectInfo(params) {
+    return new Promise((resolve, reject) => {
+      Product.findByPk(
+        params.id,
+      ).then((selectedInfo) => {
+        resolve(selectedInfo);
+      }).catch((err) => {
+        reject(err);
+      });
+    });
+  },
+  // 수정
+  update(params) {
+    return new Promise((resolve, reject) => {
+      Product.update(
+        params,
+        {
+          where: { id: params.id },
+        },
+      ).then(([updated]) => {
+        resolve({ updatedCount: updated });
+      }).catch((err) => {
+        reject(err);
+      });
+    });
+  },
+  // 삭제
+  delete(params) {
+    return new Promise((resolve, reject) => {
+      Product.destroy({
+        where: { id: params.id },
+      }).then((deleted) => {
+        resolve({ deletedCount: deleted });
+      }).catch((err) => {
+        reject(err);
+      });
+    });
+  },
+};
+
+module.exports = dao;
