+++ team_first_project01/dao/productcategoryDao.js
@@ -1,11 +1,11 @@
 const { Op } = require('sequelize');
-const { ProductReply } = require('../models/index');
+const { ProductCategory } = require('../models/index');
 
 const dao = {
   // 등록
   insert(params) {
     return new Promise((resolve, reject) => {
-      ProductReply.create(params).then((inserted) => {
+      ProductCategory.create(params).then((inserted) => {
         resolve(inserted);
       }).catch((err) => {
         reject(err);
@@ -16,10 +16,10 @@ const dao = {
   selectList(params) {
     // where 검색 조건
     const setQuery = {};
-    if (params.productNumber) {
+    if (params.categoryName) {
       setQuery.where = {
         ...setQuery.where,
-        productNumber: { [Op.like]: `%${params.productNumber}%` }, // like검색
+        categoryName: { [Op.like]: `%${params.categoryName}%` }, // like검색
       };
     }
 
@@ -27,7 +27,7 @@ const dao = {
     setQuery.order = [['id', 'DESC']];
 
     return new Promise((resolve, reject) => {
-      ProductReply.findAndCountAll({
+      ProductCategory.findAndCountAll({
         ...setQuery,
       }).then((selectedList) => {
         resolve(selectedList);
@@ -39,8 +39,8 @@ const dao = {
   // 상세정보 조회
   selectInfo(params) {
     return new Promise((resolve, reject) => {
-      ProductReply.findByPk(
-        params.productNumber,
+      ProductCategory.findByPk(
+        params.id,
       ).then((selectedInfo) => {
         resolve(selectedInfo);
       }).catch((err) => {
@@ -51,10 +51,10 @@ const dao = {
   // 수정
   update(params) {
     return new Promise((resolve, reject) => {
-      ProductReply.update(
+      ProductCategory.update(
         params,
         {
-          where: { productNumber: params.productNumber },
+          where: { id: params.id },
         },
       ).then(([updated]) => {
         resolve({ updatedCount: updated });
@@ -66,8 +66,8 @@ const dao = {
   // 삭제
   delete(params) {
     return new Promise((resolve, reject) => {
-      ProductReply.destroy({
-        where: { productNumber: params.productNumber },
+      ProductCategory.destroy({
+        where: { id: params.id },
       }).then((deleted) => {
         resolve({ deletedCount: deleted });
       }).catch((err) => {
