+++ team_first_project01/models/cart.js
@@ -1,38 +1,22 @@
 const Sequelize = require('sequelize');
 
-module.exports = class ProductReply extends Sequelize.Model {
+module.exports = class Cart extends Sequelize.Model {
   static init(sequelize) {
     return super.init({
-      replyNumber: {
+      cartNumber: {
         type: Sequelize.INTEGER,
-        primaryKey: true,
+        unique: true,
         allowNull: false,
       },
       productNumber: {
         type: Sequelize.INTEGER,
         allowNull: false,
       },
-      content: {
-        type: Sequelize.TEXT,
+      cartValue: {
+        type: Sequelize.STRING(30),
         allowNull: false,
       },
-      writerNickname: {
-        type: Sequelize.STRING(50),
-        allowNull: false,
-      },
-      reportingDate: {
-        type: Sequelize.DATE,
-        allowNull: false,
-      },
-      replyOriginNumber: {
-        type: Sequelize.INTEGER,
-        allowNull: false,
-      },
-      replyOrder: {
-        type: Sequelize.INTEGER,
-        allowNull: false,
-      },
-      replyDepth: {
+      productCount: {
         type: Sequelize.INTEGER,
         allowNull: false,
       },
@@ -47,6 +31,6 @@ module.exports = class ProductReply extends Sequelize.Model {
   }
 
   static associate(db) {
-    db.ProductReply.belongsTo(db.Product, { foreignKey: { productNumber: 'productNumber', onDelete: 'SET NULL', as: 'Product' } });
+    db.Cart.belongsTo(db.Product, { foreignKey: { productNumber: 'productNumber', onDelete: 'SET NULL', as: 'Product' } });
   }
 };
